+++ frontend/web/static/js/dashboard/shared/utils/uiRefresh.ts
@@ -8,6 +8,27 @@
 declare const debugLog: (...args: any[]) => void;
 declare const htmx: any;
 
+/**
+ * Structured logger for UI refresh operations
+ */
+const logger = {
+  info: (message: string, ...args: any[]) => {
+    if (typeof debugLog !== 'undefined') {
+      debugLog(`[UI_REFRESH] ${message}`, ...args);
+    }
+  },
+  warn: (message: string, ...args: any[]) => {
+    if (typeof debugLog !== 'undefined') {
+      debugLog(`[UI_REFRESH] ⚠️  ${message}`, ...args);
+    } else {
+      console.warn(`[UI_REFRESH] ${message}`, ...args);
+    }
+  },
+  error: (message: string, ...args: any[]) => {
+    console.error(`[UI_REFRESH] ${message}`, ...args);
+  },
+};
+
 interface RefreshConfig {
   /** Context name for logging (e.g., 'SaveFactModal', 'SavePlanModal') */
   context: string;
@@ -24,12 +45,12 @@ interface RefreshConfig {
  * @param config - Configuration for which components to refresh
  */
 export async function refreshUIAfterSave(config: RefreshConfig): Promise<void> {
-  debugLog(`[${config.context}] Refreshing UI components...`);
+  logger.info(`[${config.context}] Refreshing UI components...`);
 
   try {
     // Check if HTMX is available
     if (typeof htmx === 'undefined') {
-      console.warn('[UI_REFRESH] HTMX undefined, falling back to page reload');
+      logger.warn('HTMX undefined, falling back to page reload');
       window.location.reload();
       return;
     }
@@ -64,9 +85,9 @@ export async function refreshUIAfterSave(config: RefreshConfig): Promise<void> {
       await (window as any).reloadPlans();
     }
 
-    debugLog(`[${config.context}] UI refresh completed`);
+    logger.info(`[${config.context}] UI refresh completed`);
   } catch (error) {
-    console.error(`[${config.context}] Error refreshing UI:`, error);
+    logger.error(`[${config.context}] Error refreshing UI:`, error);
 
     // Show error toast if available
     if (typeof (window as any).showToast === 'function') {
